requirements.txt
requests
python-binance
numpy